refactor(ontology): split method.rs into folder-module
Split entity/method.rs into:
- method/extraction.rs: ExtractionMethod enum
- method/recognition.rs: RecognitionMethod enum + RecognitionMethodKind
- method/provenance.rs: PatternProvenance, ModelProvenance, ManualProvenance
- method/refinement.rs: RefinementMethod enum
